Added my main, landing, and splash screen pages
I edited the current main.dart page to match what mine was because I assume that my main page was what we needed because it would open the app. I didn't change any of the imports there, I just added the imports that I used.
I added a photo to the assets folder because I used the logo from it and without it my code wouldn't render the photo.
I didn't want to delete the already existing landing_page.dart for my landing.dart page so I left it there, whoever made that can salvage what they want from it and add it to my landing page so all of the UI changes from my landing page can function with the integration from the other landing page.
I added my splash screen page, it should automatically work with the rest of the code because I edited the main.dart file.